Fix bugs with non default hll fields
The implementation of hll fields with non default values was broken
as it passed the hll kwargs down to the Field without removing them.

Note there is a wider batch of work to support non default values in
the library as new hll fields created, such as HllEmpty need to match
the settings of the field being written to. I have not attempted to
address this, and have simply used the hll_set_defaults to control
things in the tests.